Added chakraui; The navigation menu has been moved to a separate file; dependency issues have been fixed.
