+++ src/Equipment.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import { Text, Box, useColorModeValue, Center } from '@chakra-ui/react';
 
 export default function Equipment() {
